requirements.txt
# Core Telegram Bot Framework
python-telegram-bot[all]==21.1

# Async & HTTP Client (consolidated)
httpx>=0.27.0
aiohttp==3.9.1

# Database (sqlite3 is built-in to Python)
aiosqlite==0.19.0

# AI/ML - OpenRouter as primary provider
openai==1.3.0  # For compatibility with OpenRouter API

# Configuration & Validation
python-dotenv==1.0.0
pydantic==2.5.0

# Utilities & Image Processing
requests==2.31.0
pillow==10.1.0
numpy==1.26.2
pandas==2.1.3

# Logging
colorlog==6.8.0
structlog==23.2.0

# Deployment
gunicorn==21.2.0
uvicorn==0.24.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1

# Code Quality (optional, for development)
black==23.12.0
flake8==6.1.0
pylint==3.0.3
mypy==1.7.0
